+++ src/splat4j/output/GraphGenerator.java
@@ -171,14 +171,14 @@ public class GraphGenerator {
             referenceFile = new File("reference.gp", "wb");
             FileWriter refWri = new FileWriter(referenceFile);
             for (x = 1; x < path.getLength() - 1; x++) {
-                remote = new Site("", "", path.getLat(x), path.getLon(x), 0.0);
+                remote = new Site("", "", path.getLat(x), path.getLon(x), 0.0f);
 //		remote.lat=path.lat[x];
 //		remote.lon=path.lon[x];
 //		remote.getAlt()=0.0;
                 angle = Utils.elevationAngle(destination, remote, splat, config);
 
                 if (splat.getClutter() > 0.0) {
-                    remote2 = new Site("", "", path.getLat(x), path.getLon(x), path.getElevation()[x] != 0.0 ? splat.getClutter() : 0.0);
+                    remote2 = new Site("", "", path.getLat(x), path.getLon(x), (float)(path.getElevation()[x] != 0.0 ? splat.getClutter() : 0.0));
 //			remote2.lat=path.lat[x];
 //			remote2.lon=path.lon[x];
 //
@@ -358,7 +358,7 @@ public class GraphGenerator {
                 fresnel6File  = new File("fresnel_pt_6.gp");
                 fres6Wri = new FileWriter(fresnel6File);
             }   for (x = 0; x < path.getLength() - 1; x++) {
-                remote = new Site("", "", path.getLat(x), path.getLon(x), 0.0);
+                remote = new Site("", "", path.getLat(x), path.getLon(x), 0.0f);
 //            remote.lat = path.lat[x];
 //            remote.lon = path.lon[x];
 //            remote.getAlt() = 0.0;
